+++ src/error.c
@@ -14,18 +14,18 @@
 //
 // Author: jdtang@google.com (Jonathan Tang)
 
-#include "error.h"
+#include "gumbo/error.h"
 
 #include <assert.h>
 #include <stdarg.h>
 #include <stdio.h>
 #include <string.h>
 
-#include "gumbo.h"
-#include "parser.h"
-#include "string_buffer.h"
-#include "util.h"
-#include "vector.h"
+#include "gumbo/gumbo.h"
+#include "gumbo/parser.h"
+#include "gumbo/string_buffer.h"
+#include "gumbo/util.h"
+#include "gumbo/vector.h"
 
 // Prints a formatted message to a StringBuffer.  This automatically resizes the
 // StringBuffer as necessary to fit the message.  Returns the number of bytes
