+++ src/string_buffer.c
@@ -14,15 +14,15 @@
 //
 // Author: jdtang@google.com (Jonathan Tang)
 
-#include "string_buffer.h"
+#include "gumbo/string_buffer.h"
 
 #include <assert.h>
 #include <stdlib.h>
 #include <string.h>
 #include <strings.h>
 
-#include "string_piece.h"
-#include "util.h"
+#include "gumbo/string_piece.h"
+#include "gumbo/util.h"
 
 struct GumboInternalParser;
